FifthMilestone/sender/requirements.txt
Flask
Flask-Session
gunicorn
python-dotenv
redis
bcrypt
PyJWT
requests
six
authlib
pika